third_party/WebKit/LayoutTests/inspector/sources/debugger/network-uisourcecode-provider.html

<html>
<head>
<script src="../../../http/tests/inspector/inspector-test.js"></script>
<script src="../../../http/tests/inspector/debugger-test.js"></script>
<script src="../../../http/tests/inspector/workspace-test.js"></script>
<script>
function test()
{
    var mockContentsMap = {};
    var lastResourceId = 0;

    var mockStyleSheetHeader = {
        styleSheetId: "1",
        sourceURL: "foo.css",
        sourceMapURL: "",
        origin: "regular",
        title: "",
        disabled: false
    };

    NetworkAgent.getResponseBody = function(requestId, callback)
    {
        callback(undefined, mockContentsMap[requestId], false);
    }

    PageAgent.getResourceContent = function(frameId, url, callback)
    {
        callback(undefined, mockContentsMap[url], false);
    }

    function createResourceMock(type, content)
    {
        var documentURL = "http://fake.url";
        var resourceId = ++lastResourceId + "";
        var url = documentURL + "/" + resourceId;
        var frameId = "frame-id";
        var loaderId = "loader-id";
        var mimeType;
        switch (type) {
        case WebInspector.resourceTypes.Document:
            mimeType = "text/html";
            break;
        case WebInspector.resourceTypes.Script:
            mimeType = "text/javascript";
            break;
        case WebInspector.resourceTypes.Stylesheet:
            mimeType = "text/css";
            break;
        }

        var resource = new WebInspector.Resource(WebInspector.resourceTreeModel.target(), null, url, documentURL, frameId, loaderId, type, mimeType);
        mockContentsMap[url] = content;
        WebInspector.resourceTreeModel.dispatchEventToListeners(WebInspector.ResourceTreeModel.EventTypes.ResourceAdded, resource);

        return resource;
    }

    function createScriptMock(content)
    {
        var documentURL = "http://fake.url";
        var resourceId = ++lastResourceId + "";
        var url = documentURL + "/" + resourceId;
        var script = InspectorTest.createScriptMock(url, 0, 0, false, content);
        WebInspector.debuggerModel.dispatchEventToListeners(WebInspector.DebuggerModel.Events.ParsedScriptSource, script);
    }

    function finishResource(resource)
    {
        resource.request.finished = true;
        resource.request.dispatchEventToListeners(WebInspector.NetworkRequest.Events.FinishedLoading, resource.request);
    }

    function createNetworkUISourceCodeProvider()
    {
        InspectorTest.createWorkspace();
        var target = WebInspector.targetManager.mainTarget();
        target.resourceTreeModel = new WebInspector.SDKModel(WebInspector.ResourceTreeModel, target);
        target.resourceTreeModel.inspectedPageURL = function() { return "foo"; };
        target.resourceTreeModel.inspectedPageDomain = function() { return "foo"; };
        target.debuggerModel = new WebInspector.SDKModel(WebInspector.DebuggerModel, target);
        target.debuggerModel._registerScript = function() {};
        target.cssModel = new WebInspector.CSSStyleModel(target);
        WebInspector.resourceTreeModel = target.resourceTreeModel;
        WebInspector.debuggerModel = target.debuggerModel;
        return new WebInspector.NetworkUISourceCodeProvider(InspectorTest.testNetworkWorkspaceBinding, InspectorTest.testWorkspace);
    }

    InspectorTest.runTestSuite([
        function testDocumentResource(next)
        {
            var networkUISourceCodeProvider = createNetworkUISourceCodeProvider();
            InspectorTest.addResult("Creating resource.");
            InspectorTest.waitForWorkspaceUISourceCodeAddedEvent(uiSourceCodeAdded);
            createResourceMock(WebInspector.resourceTypes.Document, "<document resource content>");

            function uiSourceCodeAdded(uiSourceCode)
            {
                InspectorTest.dumpUISourceCode(uiSourceCode, next);
            }
        },

        function testScriptResource(next)
        {
            var networkUISourceCodeProvider = createNetworkUISourceCodeProvider();
            InspectorTest.addResult("Creating resource.");
            InspectorTest.waitForWorkspaceUISourceCodeAddedEvent(uiSourceCodeAdded);
            createResourceMock(WebInspector.resourceTypes.Script, "<script resource content>");

            function uiSourceCodeAdded(uiSourceCode)
            {
                InspectorTest.dumpUISourceCode(uiSourceCode, next);
            }
        },

        function testVMScript(next)
        {
            var networkUISourceCodeProvider = createNetworkUISourceCodeProvider();
            InspectorTest.addResult("Creating script.");
            InspectorTest.waitForWorkspaceUISourceCodeAddedEvent(uiSourceCodeAdded);
            createScriptMock("<script content>");

            function uiSourceCodeAdded(uiSourceCode)
            {
                InspectorTest.dumpUISourceCode(uiSourceCode, next);
            }
        },

        function testStylesheetResource(next)
        {
            var networkUISourceCodeProvider = createNetworkUISourceCodeProvider();
            InspectorTest.addResult("Creating resource.");
            InspectorTest.waitForWorkspaceUISourceCodeAddedEvent(uiSourceCodeAdded);
            createResourceMock(WebInspector.resourceTypes.Stylesheet, "<stylesheet resource content>");

            function uiSourceCodeAdded(uiSourceCode)
            {
                InspectorTest.dumpUISourceCode(uiSourceCode, next);
            }
        },

        function testRemoveStyleSheetFromModel(next)
        {
            var networkUISourceCodeProvider = createNetworkUISourceCodeProvider();
            var target = WebInspector.targetManager.mainTarget();
            InspectorTest.waitForWorkspaceUISourceCodeAddedEvent(uiSourceCodeAdded);
            target.cssModel._styleSheetAdded(mockStyleSheetHeader);

            function uiSourceCodeAdded(uiSourceCode)
            {
                InspectorTest.addResult("Added uiSourceCode: " + InspectorTest.uiSourceCodeURL(uiSourceCode));
                InspectorTest.waitForWorkspaceUISourceCodeRemovedEvent(uiSourceCodeRemoved);
                target.cssModel._styleSheetRemoved(mockStyleSheetHeader.styleSheetId);
            }

            function uiSourceCodeRemoved(uiSourceCode)
            {
                InspectorTest.addResult("Removed uiSourceCode: " + InspectorTest.uiSourceCodeURL(uiSourceCode));
                next();
            }
        }
    ]);
};
</script>
</head>
<body onload="runTest()">
<p>Tests NetworkUISourceCodeProvider class.</p>
</body>
</html>
